Fix sidebar bugs
